Improved list_from_csv module. It is now formatted better with cleaner, more readable code.
Added jsonresponse module. This module is able to take a JSON response object of the form returned by the Google Calendar API. This response is the result of calling the Google calendar API on a time range of 12-5pm and 2 separate emails. It determines when there is a 30 minute segment of free time for these people to meet. POTENTIAL BUGS: This currently assumes that both people will have busy events during this time interval. Will likely crash if either person has no events on that calendar for that time.

Still to do: Create module that will actually get the json response, and be able to iterate over several days of a week if no open spots are found.